Move paths_on_pythonpath out of the command.test module.
